Refactor a lot
