+++ src/main.cpp
@@ -1,8 +1,8 @@
 #include "uci_interpreter/uci_interpreter.h"
 
-int main (int argc, char *argv[]) {
+int main(i32 argc, char *argv[]) {
 	initialize_engine();
-	if (argc > 1 && std::string(argv[1]) == "bench") {
+	if (argc > 1 && string{argv[1]} == "bench") {
 		bench();
 		return 0;
 	}
