+++ src/move_search/reductions.cpp
@@ -7,10 +7,10 @@
 #include "../move_gen/types/move.h"
 #include <algorithm>
 
-int lmr_reduction(const bool &pv_node, const int &ply, const bool &in_check, const bool improving,
-				  const int &move_idx, const int &depth, const Move &legal_move) {
-	int reduction = 0;
-	int lmr_depth = (pv_node || ply == 0) ? 5 : 3;
+i32 lmr_reduction(bool pv_node, i32 ply, bool in_check, bool improving, i32 move_idx,
+				  i32 depth, Move legal_move) {
+	i32 reduction = 0;
+	i32 lmr_depth = (pv_node || ply == 0) ? 5 : 3;
 	if (depth >= 3 && move_idx > lmr_depth && !in_check) {
 		reduction = static_cast<int>(lmr_table[legal_move.is_quiet()][depth][move_idx]);
 		reduction -= pv_node;
@@ -20,7 +20,7 @@ int lmr_reduction(const bool &pv_node, const int &ply, const bool &in_check, con
 	return reduction;
 }
 
-int nmp_reduction(const int& depth, const int& beta, const int& static_eval) {
-	int reduction = 3 + depth / 3 + std::min((static_eval - beta) / 200, 3);
+i32 nmp_reduction(i32 depth, i32 beta, i32 static_eval) {
+	auto reduction = 3 + depth / 3 + std::min((static_eval - beta) / 200, 3);
 	return std::max(reduction, 3);
 }
