+++ src/move_search/search_constants.h
@@ -1,8 +1,10 @@
 #pragma once
-constexpr int POS_INF_CHESS = 1000000;
-constexpr int NEG_INF_CHESS = -POS_INF_CHESS;
-constexpr int MATE_SCORE = POS_INF_CHESS / 10;
-constexpr short MAX_DEPTH = 100;
-constexpr short MAX_PLY = MAX_DEPTH;
-constexpr int MATE_BOUND = MATE_SCORE - MAX_DEPTH;
-constexpr int DEFAULT_SEARCH_TIME = 1000;
+#include "../types.h"
+
+constexpr i32 POS_INF_CHESS = 1000000;
+constexpr i32 NEG_INF_CHESS = -POS_INF_CHESS;
+constexpr i32 MATE_SCORE = POS_INF_CHESS / 10;
+constexpr i16 MAX_DEPTH = 100;
+constexpr i16 MAX_PLY = MAX_DEPTH;
+constexpr i32 MATE_BOUND = MATE_SCORE - MAX_DEPTH;
+constexpr i32 DEFAULT_SEARCH_TIME = 86'400'000;
