+++ src/move_search/tables/transposition_table.h
@@ -12,7 +12,7 @@ enum TranspositionTableEntryNodeType {
 struct TranspositionTableEntry {
 	ZobristHash zobrist_hash;
 	short depth;
-	int value;
+	i32 value;
 	TranspositionTableEntryNodeType node_type;
 	Move best_move;
 };
@@ -28,23 +28,23 @@ constexpr bool QSEARCH_TT_PV_NODE = false;
 // More to learn from here: https://github.com/kobolabs/stockfish/blob/master/tt.cpp
 class TranspositionTable {
 public:
-	TranspositionTableEntryNodeType get_node_type(const int &alpha_initial, const int &beta, const int &value);
-	int correct_mate_for_storage(int score, int ply);
-	int correct_mate_for_retrieval(int score, int ply);
-	void put(ZobristHash hash, short depth, int score, int ply, Move best_move, bool pv_node,
+	TranspositionTableEntryNodeType get_node_type(i32 alpha_initial, i32 beta, i32 value);
+	i32 correct_mate_for_storage(i32 score, i32 ply);
+	i32 correct_mate_for_retrieval(i32 score, i32 ply);
+	void put(ZobristHash hash, i16 depth, i32 score, i32 ply, Move best_move, bool pv_node,
 			 TranspositionTableEntryNodeType node_type);
 	void reset_table();
-	void resize(int mb);
+	void resize(i32 mb);
 	void prefetch(ZobristHash hash);
 	TranspositionTableSearchResults probe_for_move_ordering(ZobristHash hash);
-	TranspositionTableSearchResults probe_for_search(ZobristHash hash, int depth, int ply);
-	TranspositionTableSearchResults probe_eval(ZobristHash hash, int ply);
-	int mb_to_entries(int mb);
-	explicit TranspositionTable(int mb = 64);
+	TranspositionTableSearchResults probe_for_search(ZobristHash hash, i32 depth, i32 ply);
+	TranspositionTableSearchResults probe_eval(ZobristHash hash, i32 ply);
+	i32 mb_to_entries(i32 mb);
+	explicit TranspositionTable(i32 mb = 64);
 	size_t entry_count();
 
 private:
-	uint64_t get_index(uint64_t zobrist_hash);
+	u64 get_index(u64 zobrist_hash);
 	std::vector<TranspositionTableEntry> transposition_table;
 };
