+++ src/move_search/types.cpp
@@ -6,7 +6,7 @@
 #include "../move_gen/types/move.h"
 
 std::ostream& operator<<(std::ostream& os, const Line& line) {
-	for(const Move & i : line) {
+	for(const Move& i : line) {
 		if (i == Move()) break;
 		os << i << " ";
 	}
