+++ src/move_search/types.h
@@ -6,15 +6,15 @@
 
 struct ScoredMove {
 	Move move;
-	int score{};
+	i32 score{};
 };
 
 using ScoredMoves = std::vector<ScoredMove>;
 
 struct BestMoveSearchParameters {
-	short depth = MAX_DEPTH;
-	int hard_time_limit = DEFAULT_SEARCH_TIME;
-	int soft_time_limit = DEFAULT_SEARCH_TIME;
+	i16 depth = MAX_DEPTH;
+	i32 hard_time_limit = DEFAULT_SEARCH_TIME;
+	i32 soft_time_limit = DEFAULT_SEARCH_TIME;
 	bool debug_info = false;
 };
 
@@ -25,16 +25,16 @@ struct PVSData {
 	Move best_move;
 	Move final_best_move;
 	bool search_completed{};
-	int value{};
+	i32 value{};
 	// triangular-table-table
 	PV pv{};
-	uint64_t nodes_searched{};
-	int seldepth{};
+	u64 nodes_searched{};
+	i32 seldepth{};
 
 	Move moves_made[MAX_PLY];
-	int time_limit{};
+	i32 time_limit{};
 
 	Move excluded_moves[MAX_PLY]{};
-	uint64_t nodes_spend[NSQUARES][NSQUARES]{};
-	int evals[MAX_PLY]{};
+	u64 nodes_spend[NSQUARES][NSQUARES]{};
+	i32 evals[MAX_PLY]{};
 };
