+++ src/uci_interpreter/bench_fens.h
@@ -1,7 +1,7 @@
 #pragma once
 #include "iostream"
 
-constexpr int BENCH_SIZE = 50;
+constexpr i32 BENCH_SIZE = 50;
 constexpr const char *BENCH_FENS[BENCH_SIZE] = {
 		"r3k2r/2pb1ppp/2pp1q2/p7/1nP1B3/1P2P3/P2N1PPP/R2QK2R w KQkq a6 0 14",
 		"4rrk1/2p1b1p1/p1p3q1/4p3/2P2n1p/1P1NR2P/PB3PP1/3R1QK1 b - - 2 24",
