+++ src/utils/fen_constants.h
@@ -3,10 +3,6 @@
 //
 #pragma once
 
-constexpr auto UCI_DEBUG_STRING = "info string ";
-
 constexpr auto INITIAL_BOARD_FEN = "rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR w KQkq - 0 1";
 constexpr auto KIWIPETE_FEN = "r3k2r/p1ppqpb1/bn2pnp1/3PN3/1p2P3/2N2Q1p/PPPBBPPP/R3K2R w KQkq - 0 1";
 constexpr auto TALKCHESS_FEN = "rnbq1k1r/pp1Pbppp/2p5/8/2B5/8/PPP1NnPP/RNBQK2R w KQ - 1 8";
-constexpr auto PERFT_RESULTS_POS4_FEN = "r3k2r/Pppp1ppp/1b3nbN/nP6/BBP1P3/q4N2/Pp1P2PP/R2Q1RK1 w kq - 0 1";
-constexpr auto ASSERTION_ERR_SEARCH_DEPTH = "Test not completed, did not search deep enough";
