+++ src/utils/helpers.cpp
@@ -4,17 +4,17 @@
 
 #include "helpers.h"
 
-std::vector<std::string> split(const std::string& s, const std::string& delimiter) {
+std::vector<string> split(const string& s, const string& delimiter) {
 	size_t pos_start = 0, pos_end, delim_len = delimiter.length();
-	std::string token;
-	std::vector<std::string> res;
+	string token;
+	std::vector<string> res;
 
-	while ((pos_end = s.find (delimiter, pos_start)) != std::string::npos) {
-		token = s.substr (pos_start, pos_end - pos_start);
+	while ((pos_end = s.find (delimiter, pos_start)) != string::npos) {
+		token = s.substr(pos_start, pos_end - pos_start);
 		pos_start = pos_end + delim_len;
-		res.push_back (token);
+		res.push_back(token);
 	}
 
-	res.push_back(s.substr (pos_start));
+	res.push_back(s.substr(pos_start));
 	return res;
 }
